% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cn}
\alias{cn}
\title{show column names}
\usage{
cn(df)
}
\arguments{
\item{x}{= a data.frame}
}
\value{
a matrix with `data.frame` names
}
\description{
A convenience function to show column names
}
\examples{
dat = data.frame(matrix(rnorm(100), ncol=10))
names(dat) = letters[1:length(dat)]
cn(dat)
}
\concept{utilities}
