% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychometrics.R
\name{simulate_scores}
\alias{simulate_scores}
\title{Simulate scores with given reliability}
\usage{
simulate_scores(n, rel, mu_true = 0, sd_true = 1)
}
\arguments{
\item{n}{= number of observation
rel = reliability value (from 0 to 1)
mu_true = mean of population
sd_true = standard deviation of the population}
}
\value{
a data frame with two columns, `true.s` and `obs.s`.
}
\description{
A function to simulate scores providing reliability
}
