% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychometrics.R
\name{simulate_obs}
\alias{simulate_obs}
\title{plot simulated scores and optional a cut-off}
\usage{
simulate_obs(
  Ts,
  n.obs,
  Rel = 0.7,
  mu_true,
  sd_true,
  label = "True - observed relationship",
  plot_thresh = FALSE
)
}
\arguments{
\item{n.obs}{= number of observations
Ts = the current True score (on which observed are computed)
Rel = reliability value (from 0 to 1)
mu_true = mean of population
sd_true = standard deviation of the population
label = the title of the plot
plot_thresh = `logical`, should the threshold be plotted?}
}
\value{
a plot with the simulated scores.
}
\description{
A function to simulate scores providing reliability
}
