% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychometrics.R
\name{simulate_obs}
\alias{simulate_obs}
\title{plot simulated scores and optional a cut-off}
\usage{
simulate_obs(
  Ts,
  n,
  rel = 0.7,
  mu_true,
  sd_true,
  label = "True - observed relationship",
  plot_res = TRUE,
  plot_thresh = FALSE
)
}
\arguments{
\item{n}{= number of observations
Ts = the current True score (on which observed are computed)
rel = reliability value (from 0 to 1)
mu_true = mean of population
sd_true = standard deviation of the population
label = the title of the plot
plot_res = `logical`, a plot with the results.
plot_thresh = `logical`, should the threshold be plotted?}
}
\value{
a vector with observed scores given the true. Optiomally, plot with the simulated scores.
}
\description{
A function to simulate scores providing reliability
}
\examples{
simulate_obs(Ts = 5, n = 20, rel=0.8, mu_true = 8, sd_true = 2)
}
\seealso{
Other psychometrics: 
\code{\link{simulate_scores}()}
}
\concept{psychometrics}
